Defend tests against rogue workers
When a test fails due to an assertion failure the execution of the rest
of the test is halted. This means that the shutdown method is not called
on the worker. Thus, all subsequent tests will have two workers that are
polling the database and posting jobs. This can lead to multiple
subsequent failures of completely unrelated tests, which is undesirable.

Fix this situation by providing tests with an environment with a worker
where the shutdown is ensured even in the case of a test failure,
isolating the unit tests from each other.