Respect queues when collecting jobs to execute
If a queue has no jobs running the poller would post as many jobs from
that queue as possible, according to the priority. This is of course
unacceptable, as "queue" implies sequential processing.

Fix this by refactoring the queued_db_jobs method. The procedure now is
as follows:

  1. Lock all queues
  2. For each queue which may be processed, select the job with the
     highest priority. If unqueued jobs may be processed, select the
     maximum amount of such jobs.
  3. Combine the results of the single selects from step 2.
  4. Order the results and limit to the correct number of jobs.
  5. Return this list of jobs.

Step 3 requires some Arel acrobatics and literal SQL building, as there
is no way of building the multiple-union query in a way that generates
proper SQL.

Fix #13