Fix sequencing jobs of same queue
Due to a bug in the poller it was possible that a job of a queue gets
scheduled even if there is a job already running in the same queue. This
was due to a wrong keyword ('running' instead of 'started') in the
polling query that determines 'bad queues'. This has now been fixed
properly by using the respective constants of the DbJob model.

This commit also includes updated unit tests that reproduce this problem
and prove the fix.