Fix union of CLOB fields for Oracle
When valid (i.e. runnable) jobs are present in multiple queues, the
poller builds a query using UNION. Prior to this change, the queries
thus unioned were of the type "SELECT *", which included a CLOB-type
field. Oracle does not know how to handle such a case.

To address this issue, the subselects that are unioned are now of the
type "SELECT id", circumventing the CLOB field entirely.

In the course of this refactoring, an additional layer is wrapped around
the query to recover the entire records from the jobs table based on the
ids yielded by the UNION query.