Fix query parts addition for the UNION query
Oracle cannot perform a union of the following sort:

  ( SELECT ... ) UNION ( SELECT ... )

Instead, it should look as follows:

  SELECT ... UNION SELECT ...

The UNION'ed parts contain ordering clauses to limit the number of jobs
to the top job per queue, which need to be scoped to the parts. That is
the reason why the parentheses were used in the first place.

The issue is addressed by wrapping the parts of the UNION query in an
additional SELECT. This is not enough, however, as now MySQL will
complain about missing table aliases. So, these are added as well,
resulting in a UNION query of the following form:

  SELECT * FROM ( SELECT ... ) union_0
  UNION SELECT * FROM ( SELECT ... ) union_1
  UNION SELECT * FROM ( SELECT ... ) union_2
  ...

This is tested and confirmed to work with MySQL and Oracle DB with
multiple jobs in multiple queues.