Always perform jobs with no queue
Due to the way the valid queues for querying the jobs table was
calculated, the nil queue is ignored once it has a job that is either
locked or started. By adding an additional condition to the query that
is responsible for selecting the list of invalid queues this problem is
successfully circumvented.

At the same time, the variable 'bad_queries_select' was renamed to the
more appropriate 'bad_queues_select'.

Two new unit tests concerning the poller are added: One for the specific
case of checking that the nil queue is still valid even if a job therein
is locked, and one that ensures no queue is marked valid when no job
exists at all.