Simplify instant repolling
Simplify instant repolling to, if enabled, always perform an instant
repoll whenever a job has finished. This removes the timestamp check
which was previously in place. The reason behind this is that whenever
a job has completed, a new slot opened in the job pool and thus a
re-poll is in order.