Fix #22
This fixes a deadlock exception that arose when calling `wait` on a pool
that has not been shut down yet. Before this commit,
ThreadPoolExecutor's `wait_for_termination` has been used which is, as
stated in the official documentation, only to be called *after* calling
shutdown on the thread pool executor.

This only seemed to cause problems when there are no other threads
beside the main thread and the pool threads around, which is why this
exception didn't always happen and why in the test supplied in this
commit we needed to forcefully exit all other threads and disconnect
from the database.

The solution in this commit is to use a loop paired with sleep and a
check for `@pool.shutdown?`. While not as elegant as the previous
solution, it should be more robust and work in any case.